LookLearn with Working Messaging System

+++ app/Http/Controllers/PostsController.php
@@ -102,6 +102,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -143,6 +149,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error', 'Unauthorized page.');
+        }
+
         $post->delete();
         return redirect('/posts')->with('success', 'Post deleted.');
     }
